+++ CHANGELOG.md
@@ -2,5 +2,5 @@
 
 ### 1.1.0
 
-    - Add benchmarks
-    - Rename `toJSON` to `toString` to prevent misunderstanding
+- Add benchmarks
+- Remove `toJSON` function as it is simply just `JSON.stringify` applied to `toObject` which is not useful